feat: make destroy() and polyfill() chainable,allow global flag override
+ added further tests
+ adjusted package entries
+ fix regression, error in node env

For more details, see CHANGELOG
